refact: 퀴즈 생성 화면 파일 이름 변경
